BFL/requirements.txt
absl-py==2.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
dm-tree==0.1.8
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
h5py==3.10.0
idna==3.6
ivon-opt==0.1
Jinja2==3.1.2
keras==3.0.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
rich==13.7.0
scipy==1.11.4
six==1.16.0
sympy==1.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0